feat(history): Add archive functionality to history list
Enables users to archive previously saved quizzes. This change adds an "Archive" button next to each quiz in the history list. Clicking this button saves the quiz to the user's archive, preventing it from being accidentally deleted and allowing it to be used as a template for new quizzes in the future. The UI is updated to include an archive icon.